Revert "[refactor/#179] SignUp 개편 및 관련 코드 수정"
Merge 방식 개편(squash -> create)에 따른 Revert

+++ module-auth/src/main/java/com/inhabas/api/auth/domain/oauth2/member/dto/UpdateRequestDto.java
@@ -11,7 +11,7 @@ import java.util.List;
 @AllArgsConstructor
 public class UpdateRequestDto {
 
-    private List<Long> memberIdList;
+    private List<Integer> memberIdList;
 
     @Schema(example = "pass, fail")
     private String state;
